+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/common/setup/BlockEntityTypes.kt
@@ -137,5 +137,14 @@ object BlockEntityTypes {
         )
     }
 
+    val NETWORK_MANAGER: Supplier<BlockEntityType<NetworkManagerBlockEntity>> = ModPlatform.registerBlockEntity(
+        modId("network_manager"),
+    ) {
+        PlatformToolkit.get().createBlockEntityType(
+            ::NetworkManagerBlockEntity,
+            Blocks.NETWORK_MANAGER.get(),
+        )
+    }
+
     fun doSomething() {}
 }