+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/data/ModLootTableProvider.kt
@@ -34,6 +34,7 @@ object ModLootTableProvider {
         lootTable.dropSelf(consumer, Blocks.INFORMATIVE_REGISTRY)
         lootTable.dropSelf(consumer, Blocks.STATUE_WORKBENCH)
         lootTable.dropSelf(consumer, Blocks.ENTITY_LINK)
+        lootTable.dropSelf(consumer, Blocks.NETWORK_MANAGER)
         lootTable.computedDrop(Blocks.FLEXIBLE_REALITY_ANCHOR)
         lootTable.computedDrop(Blocks.FLEXIBLE_STATUE)
         lootTable.validate()